+++ mobile/android/android-components/components/browser/engine-gecko/src/main/java/mozilla/components/browser/engine/gecko/GeckoEngineSession.kt
@@ -262,6 +262,9 @@ class GeckoEngineSession(
             { throwable ->
                 // Log the error. There is nothing we can do otherwise.
                 logger.error("Save to PDF failed.", throwable)
+                notifyObservers {
+                    onSaveToPdfError(throwable)
+                }
                 GeckoResult()
             },
         )