+++ mobile/android/android-components/components/browser/state/src/main/java/mozilla/components/browser/state/engine/EngineObserver.kt
@@ -411,4 +411,8 @@ internal class EngineObserver(
             ),
         )
     }
+
+    override fun onSaveToPdfError(throwable: Throwable) {
+        store.dispatch(EngineAction.SaveToPdfExceptionAction(tabId, throwable))
+    }
 }